refactor: move the setup and analysis of Bellman/DSOP problems to bellman.py
